refactor(utils): move helper functions into utils module
